tidy up and refactor
