+++ cmd/root.go
@@ -26,7 +26,6 @@ import (
 	"strings"
 
 	"github.com/rsvihladremio/dremio-diagnostic-collector/collection"
-	"github.com/rsvihladremio/dremio-diagnostic-collector/helpers"
 	"github.com/rsvihladremio/dremio-diagnostic-collector/kubernetes"
 	"github.com/rsvihladremio/dremio-diagnostic-collector/ssh"
 	"github.com/spf13/cobra"
@@ -137,11 +136,9 @@ ddc --k8s --kubectl-path /opt/bin/kubectl --coordinator default:app=dremio-coord
 			collectorStrategy = ssh.NewCmdSSHActions(sshKeyLoc, sshUser)
 		}
 		// Create ref to real file system (since with testing we redirect the argument to a mock object)
-		cfs := helpers.FileSystem{}
 		err = collection.Execute(collectorStrategy,
 			logOutput,
 			collectionArgs,
-			cfs,
 		)
 
 		if err != nil {
